feat: MVP release - simplify to core functionality
🎯 SCOPE PIVOT: From enterprise platform to focused MVP

MAJOR CHANGES:
- Replace complex main.ts with simplified MVP version
- Backup original enterprise version as main-enterprise.ts
- Focus on 4 core features: analysis, AI insights, summaries, privacy
- Remove advanced pattern detection, orchestration, and enterprise features
- Simplify settings interface and user experience

CORE FEATURES (MVP):
✅ Smart note analysis (structure, metadata, sections)
✅ AI-powered insights (OpenAI + Ollama support)
✅ Simple summary creation with backlinks
✅ Privacy protection (tags + folders)
✅ Intelligent caching and performance optimization
✅ User-friendly interface with clear settings

REMOVED COMPLEXITY:
⏸️ Advanced pattern detection engine (80% complete, deferred)
⏸️ Multi-stage AI orchestration workflows
⏸️ Vault-wide correlation analysis
⏸️ Enterprise-grade performance optimization
⏸️ Complex prompt engineering and customization
⏸️ Batch processing and advanced workflows

DOCUMENTATION:
- Comprehensive MVP README with setup and usage guide
- Detailed CHANGELOG documenting scope reduction rationale
- Project overview with reality check and publication roadmap

TECHNICAL:
- Updated manifest.json and package.json to v0.9.0
- Successful build verification
- Simplified settings interface
- Streamlined command structure
- Optimized for 2-week publication timeline

WHY THIS MATTERS:
- Transforms 6x scope creep into focused, shippable product
- Leverages 80% complete enterprise foundation
- Provides immediate user value
- Enables rapid iteration based on community feedback
- Establishes solid foundation for future enterprise features

READY FOR: Community testing, feedback, and publication
